test: get, create and update products
test all possible scenarios of get, create and update products.

+++ src/main/java/com/ksero/products/products/service/ProductServiceImpl.java
@@ -68,6 +68,10 @@ public class ProductServiceImpl implements ProductService {
 
     @Override
     public ResponseEntity<?> delete(Long id) {
+
+        if (!repository.existsById(id))
+            throw new ResourceNotFoundException(ENTITY, id);
+
         return repository.findById(id).map(
                         product -> {
                             repository.delete(product);
